Fix battery level getter
The problem in existing implementation is that the 0 value is being skipped. For example when this.stateObj.attributes.battery is 0 and battery_level attrib was not present then the "state" value is taken. 

When battery or battery_level attributes are present and they contain numeric value then such value should be returned - even when it is 0.